add VAE models to SLAC25.models
